+++ packages/agent-infra/mcp-servers/browser/src/request-context.ts
@@ -2,6 +2,7 @@
  * Copyright (c) 2025 Bytedance, Inc. and its affiliates.
  * SPDX-License-Identifier: Apache-2.0
  */
+import type { MiddlewareFunction } from 'mcp-http-server';
 import { AsyncLocalStorage } from 'async_hooks';
 import type { RequestContext } from 'mcp-http-server';
 
@@ -14,3 +15,18 @@ export const setRequestContext = (context: RequestContext) => {
 export const getRequestContext = (): RequestContext | undefined => {
   return requestContext.getStore();
 };
+
+// @deprecated: use addMiddleware instead
+export const onBeforeStart = () => {};
+
+const middlewares: MiddlewareFunction[] = [];
+export const addMiddleware = (middleware: MiddlewareFunction) => {
+  console.log('addMiddleware', middleware);
+  middlewares.push(middleware);
+};
+export const getMiddlewares = () => {
+  return middlewares;
+};
+
+export { setConfig } from './server.js';
+export { BaseLogger } from '@agent-infra/logger';