+++ packages/agent-infra/mcp-servers/browser/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @agent-infra/mcp-server-browser
 
+## 1.2.5
+
+### Patch Changes
+
+- e5f4157: fix: request-context cjs
+- fix: cjs
+
+## 1.2.5-beta.0
+
+### Patch Changes
+
+- fix: request-context cjs
+
+## 1.2.4
+
+### Patch Changes
+
+- fix(mcp-server): custom logger and middlewares timing issue
+
 ## 1.2.3
 
 ### Patch Changes